refactor: move response to internal
