+++ internal/handler/auth_handler.go
@@ -3,7 +3,7 @@ package handler
 import (
 	operror "basic_server/internal/errors"
 	"basic_server/internal/request"
-	"basic_server/response"
+	"basic_server/internal/response"
 	"basic_server/service"
 	"basic_server/utils"
 	"net/http"
