+++ server/handler/register_handler.go
@@ -5,7 +5,7 @@ import (
 
 	operror "basic_server/internal/errors"
 	"basic_server/internal/request"
-	"basic_server/response"
+	"basic_server/internal/response"
 	"basic_server/service"
 	"basic_server/utils" //nolint
